modules/platform_integration/youtube_proxy/src/youtube_proxy.py
import os
import logging
from googleapiclient.discovery import build
from google.oauth2.credentials import Credentials

class YouTubeProxy:
    """
    Acts as a unified, WSP-compliant interface for all YouTube operations.
    This class orchestrates underlying authentication and communication modules.
    """

    def __init__(self, credentials: Credentials):
        """
        Initializes the YouTubeProxy with authenticated credentials.

        :param credentials: An OAuth2 credentials object from google.oauth2.credentials.
        """
        if not credentials:
            raise ValueError("Credentials must be provided for YouTubeProxy initialization.")
        
        self.service = build('youtube', 'v3', credentials=credentials)
        self.logger = logging.getLogger(__name__)
        self.logger.info("YouTubeProxy initialized successfully.")

    def find_active_livestream(self, channel_id: str) -> (str, str):
        """
        Finds the active livestream for a given YouTube channel.
        This reconstructs the logic from the missing StreamResolver module.

        :param channel_id: The ID of the YouTube channel to search.
        :return: A tuple containing the (video_id, live_chat_id) or (None, None) if not found.
        """
        self.logger.info(f"Searching for active livestream for channel ID: {channel_id}")
        try:
            search_response = self.service.search().list(
                channelId=channel_id,
                eventType='live',
                type='video',
                part='snippet'
            ).execute()

            if not search_response.get('items'):
                self.logger.info("No active livestream found for the channel.")
                return None, None

            # Assuming the first result is the desired livestream
            first_result = search_response['items'][0]
            video_id = first_result['id']['videoId']
            live_chat_id = first_result['snippet']['liveChatId']
            
            self.logger.info(f"Found active livestream. Video ID: {video_id}, Chat ID: {live_chat_id}")
            return video_id, live_chat_id

        except Exception as e:
            self.logger.error(f"An error occurred while searching for livestream: {e}")
            return None, None

    def get_stream_title(self, video_id: str) -> str:
        """
        Retrieves the title for a given video ID.

        :param video_id: The ID of the YouTube video.
        :return: The video title as a string, or "Unknown Stream" if not found.
        """
        self.logger.info(f"Retrieving title for video ID: {video_id}")
        try:
            video_response = self.service.videos().list(
                id=video_id,
                part='snippet'
            ).execute()

            if not video_response.get('items'):
                self.logger.warning(f"Could not find video with ID: {video_id}")
                return "Unknown Stream"

            title = video_response['items'][0]['snippet']['title']
            self.logger.info(f"Found title: '{title}'")
            return title

        except Exception as e:
            self.logger.error(f"An error occurred while retrieving video title: {e}")
            return "Unknown Stream"

    # ... Other methods for chat, etc., will be added here. 