modules/platform_integration/youtube_proxy/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib 